+++ rqueue-core/src/main/java/com/github/sonus21/rqueue/config/RqueueListenerBaseConfig.java
@@ -111,7 +111,6 @@ public abstract class RqueueListenerBaseConfig {
         connectionFactory,
         simpleRqueueListenerContainerFactory.getReactiveRedisConnectionFactory(),
         sharedConnection,
-        reactiveEnabled,
         version);
   }