+++ rqueue-core/src/test/java/com/github/sonus21/rqueue/config/RqueueListenerBaseConfigTest.java
@@ -168,7 +168,7 @@ class RqueueListenerBaseConfigTest extends TestBase {
   void getMessageTemplateUseFromFactory() throws IllegalAccessException {
     SimpleRqueueListenerContainerFactory factory = new SimpleRqueueListenerContainerFactory();
     RqueueListenerConfig rqueueListenerConfig = createConfig(factory);
-    RqueueConfig rqueueConfig = new RqueueConfig(redisConnectionFactory, null, false, false, 1);
+    RqueueConfig rqueueConfig = new RqueueConfig(redisConnectionFactory, null, false, 1);
     factory.setRqueueMessageTemplate(rqueueMessageTemplate);
     assertEquals(rqueueMessageTemplate, rqueueListenerConfig.getMessageTemplate(rqueueConfig));
   }
@@ -177,7 +177,7 @@ class RqueueListenerBaseConfigTest extends TestBase {
   void getMessageTemplateSetTemplateInFactory() throws IllegalAccessException {
     SimpleRqueueListenerContainerFactory factory = new SimpleRqueueListenerContainerFactory();
     RqueueListenerConfig rqueueListenerConfig = createConfig(factory);
-    RqueueConfig rqueueConfig = new RqueueConfig(redisConnectionFactory, null, false, false, 1);
+    RqueueConfig rqueueConfig = new RqueueConfig(redisConnectionFactory, null, false, 1);
     RqueueMessageTemplate template = rqueueListenerConfig.getMessageTemplate(rqueueConfig);
     assertNotNull(template);
     assertEquals(template, factory.getRqueueMessageTemplate());